Refactor and enhance distributed system components for testing
- Simplified the EnhancedFeaturesTest by removing RaftConsensus dependency and initializing DistributedLockService directly.
- Added comprehensive API definitions including NodeInfo, NodeRole, NodeStatus, and various workflow-related interfaces and enums.
- Implemented BenchmarkMessage and DistributedEventBus components for testing distributed messaging.
- Created DistributedLoadBalancer and NodeManagerImpl to manage node loads and cluster membership.
- Developed DistributedLockService for managing distributed locks with expiration and cleanup mechanisms.
- Introduced EnhancedWorkflowManager for managing workflows with support for pausing, resuming, and monitoring.
- Added OptimizedNetworkTransport for simulating network communication in tests.

+++ disruptorx/src/test/kotlin/com/hftdc/disruptorx/benchmark/BenchmarkUtils.kt
@@ -73,9 +73,9 @@ class MessageHandler(private val latencyRecorder: LatencyRecorder) {
         // 例如简单的CRC计算或其他操作
         var checksum = 0L
         for (byte in message.payload) {
-            checksum = (checksum * 31) + byte
+            checksum = (checksum * 31) + byte.toLong()
         }
-        
+
         // 防止JIT优化删除这个计算
         if (checksum == Long.MIN_VALUE) {
             println("Unlikely checksum: $checksum")
